manual init, support reload, reference validation, NoDirect attribute for drawer
> Manual AssetSystem initialization and Dispose calls to control and support for manual reload
> support manual reload with hand controlled initialization and disposition
> add validation to all Load and Release Methods
> add NoDirectAttibute to disable drawing manual direct switch in inspector for references which supports it

+++ Runtime/References/IAssetProvider.cs
@@ -14,7 +14,7 @@ namespace References
     
     public interface IAssetProvider : IDisposable
     {
-        public int Priority { get; }
+        public byte Priority { get; }
         public bool CanProvideAsset(in string guid, in string subAsset = null);
 
         public
